+++ trans/src/80x86/dw2_config.c
@@ -117,7 +117,7 @@ void
 out_dwf_labdiff(long lo, long hi)
 {
   if (hi == lo) {
-    outn((long)0);
+    outn(0);
   } else {
     out_dwf_label(hi, 0);
     outs(" - ");
